mpu6050: dtsi: Add I2C device description for mpu6050
Mpu6050 is an i2c device and will be connected
to the i2c2 port on Orange Pi One.

So its description should be added to the i2c2 section
which is in am335x-bone-common.dtsi
and contains the following information (minimal) :

1. compatible = "gl,mpu6050";
   This is a driver match string, i2c driver should contain the same string
   in the id_table.

2. reg = <0x68>;
   Device i2c address. Includes r/w bit.

3. status = "okay";
   Not okay devices won't be loaded.

Also example of a patch with dtsi changes added. It can be applied during armbian compiation.

Signed-off-by: Oleksii.Honchar <oleksii.honchar@globallogic.com>